Fix: pageNotFound & change Routes w/Sidebar to notAuthenticated
