+++ apps/frontend/src/components/ui/NotFoundIllustration.tsx
@@ -1,0 +1,11 @@
+export default function NotFoundIllustration() {
+  return (
+    <div>
+      <img
+        src="/illustrations/file_search.svg"
+        alt="Site not found illustration where files are searched but not found"
+        className="h-64 w-64 object-contain"
+      />
+    </div>
+  );
+}
