+++ apps/frontend/src/routes/__root.tsx
@@ -1,27 +1,34 @@
-import { createRootRoute, Outlet, Link } from '@tanstack/react-router';
+import { createRootRoute, Outlet } from '@tanstack/react-router';
+import Sidebar from '../components/sidebar/sidebar';
+import NotFoundView from '../views/NotFoundView';
 
-function NotFound() {
+export const Route = createRootRoute({
+  component: AuthenticatedLayout,
+  notFoundComponent: NotFoundWithSidebar,
+});
+
+function AuthenticatedLayout() {
   return (
-    <div className="mx-auto max-w-xl p-8 text-center">
-      <h1 className="text-2xl font-semibold">Seite nicht gefunden</h1>
-      <p className="mt-2 text-[rgb(var(--muted-foreground-rgb))]">
-        Die angeforderte Seite existiert nicht.
-      </p>
-      <Link
-        to="/"
-        className="mt-6 inline-block rounded-xl bg-[var(--primary)] px-4 py-2 font-medium text-white hover:bg-[var(--primary-hover)]"
-      >
-        Zur Startseite
-      </Link>
+    <div className="min-h-screen">
+      <Sidebar />
+      <main className="relative ml-16">
+        <div className="p-6">
+          <Outlet />
+        </div>
+      </main>
     </div>
   );
 }
 
-export const Route = createRootRoute({
-  component: () => (
-    <div className="min-h-screen bg-[var(--bg)] text-[var(--fg)]">
-      <Outlet />
+function NotFoundWithSidebar() {
+  return (
+    <div className="min-h-screen">
+      <Sidebar />
+      <main className="relative ml-16">
+        <div className="p-6">
+          <NotFoundView />
+        </div>
+      </main>
     </div>
-  ),
-  notFoundComponent: NotFound,
-});
+  );
+}
