+++ apps/frontend/src/routes/_authenticated.tsx
@@ -1,19 +1,3 @@
-import { createFileRoute, Outlet } from '@tanstack/react-router';
-import Sidebar from '../components/sidebar/sidebar';
+import { createFileRoute } from '@tanstack/react-router';
 
-export const Route = createFileRoute('/_authenticated')({
-  component: AuthenticatedLayout,
-});
-
-function AuthenticatedLayout() {
-  return (
-    <div className="min-h-screen">
-      <Sidebar />
-      <main className="relative ml-16">
-        <div className="p-6">
-          <Outlet />
-        </div>
-      </main>
-    </div>
-  );
-}
+export const Route = createFileRoute('/_authenticated')({});
