+++ apps/frontend/src/routes/_authenticated/login.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import LoginView from '../views/LoginView';
+import LoginView from '../../views/LoginView';
 
 export const Route = createFileRoute('/login')({
   component: LoginView,
