+++ apps/frontend/src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute, redirect } from '@tanstack/react-router';
-import HomeView from '../../views/HomeView';
+import HomeView from '../views/HomeView';
 
 export const Route = createFileRoute('/_authenticated/')({
   beforeLoad: () => {
