+++ apps/frontend/src/routes/profile.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router';
-import ProfileView from '../../views/ProfileView';
+import ProfileView from '../views/ProfileView';
 
 export const Route = createFileRoute('/_authenticated/profile')({
   component: ProfileView,
