+++ package.json
@@ -44,5 +44,8 @@
     "prettier": "^3.6.2",
     "typescript": "^5.9.2",
     "typescript-eslint": "^8.45.0"
+  },
+  "dependencies": {
+    "eslint-plugin-react": "^7.37.5"
   }
 }
